Finished register pair set/get. Tested INC (no carry), 16 bit LD.

+++ src/GameBoy/Instructions.java
@@ -35,6 +35,6 @@ interface Operation {
      * @param memory Memory(RAM+ROM) to read/write from
      * @param args   additional arguments for any operations
      */
-    void cmd(Registers regs, Memory memory, int[] args);
+    void cmd(Registers regs, Memory memory, short[] args);
 
 }
